Abstract analytics client
We should enforce analytics being collected. Split out the
implementation to send analytics into an implementation module called
forgerock-openbanking-analytics-webclient.

This means we will need to add this dependency into the services in
ob-reference-implementation.
