+++ forgerock-openbanking-analytics-client/src/main/java/com/forgerock/openbanking/analytics/configuration/MetricsConfiguration.java
@@ -20,25 +20,130 @@
  */
 package com.forgerock.openbanking.analytics.configuration;
 
+import com.forgerock.openbanking.analytics.model.entries.EndpointUsageEntry;
+import com.forgerock.openbanking.analytics.model.entries.JwtsGenerationEntry;
+import com.forgerock.openbanking.analytics.model.entries.JwtsValidationEntry;
+import com.forgerock.openbanking.analytics.model.entries.TokenUsage;
+import com.forgerock.openbanking.analytics.services.CallBackCountersKPIService;
+import com.forgerock.openbanking.analytics.services.ConsentMetricService;
+import com.forgerock.openbanking.analytics.services.DirectoryCountersKPIService;
+import com.forgerock.openbanking.analytics.services.MetricService;
+import com.forgerock.openbanking.analytics.services.PsuCounterEntryKPIService;
+import com.forgerock.openbanking.analytics.services.SessionCountersKPIService;
+import com.forgerock.openbanking.analytics.services.TokenUsageService;
+import com.forgerock.openbanking.analytics.services.TppEntriesKPIService;
 import com.forgerock.openbanking.config.ApplicationConfiguration;
+import com.forgerock.openbanking.model.oidc.AccessTokenResponse;
 import com.nimbusds.jose.jwk.JWKSet;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Service;
+
+import java.util.Collection;
+import java.util.List;
 
 @Service
 public class MetricsConfiguration implements ApplicationConfiguration {
 
-    @Value("${metrics.endpoints.root}")
-    public String rootEndpoint;
+    @ConditionalOnMissingBean(CallBackCountersKPIService.class)
+    @Bean
+    CallBackCountersKPIService callBackCountersKPIService() {
+        return callBackCounterEntries -> {};
+    }
 
-    @Value("${metrics.endpoints.endpoint-usage.add-entries}")
-    public String endpointUsageAddEntries;
+    @ConditionalOnMissingBean(ConsentMetricService.class)
+    @Bean
+    ConsentMetricService consentMetricService() {
+        return consentMetricService -> {};
+    }
 
-    @Value("${metrics.endpoints.jwts-generation.add-entries}")
-    public String jwtsGenerationAddEntries;
+    @ConditionalOnMissingBean(DirectoryCountersKPIService.class)
+    @Bean
+    DirectoryCountersKPIService directoryCountersKPIService() {
+        return directoryCountersKPIService -> {};
+    }
 
-    @Value("${metrics.endpoints.jwts-validation.add-entries}")
-    public String jwtsValidationAddEntries;
+    @ConditionalOnMissingBean(MetricService.class)
+    @Bean
+    MetricService metricService() {
+        return new MetricService() {
+            @Override
+            public void addEndpointUsageEntry(EndpointUsageEntry endpointUsageEntry) {
+
+            }
+
+            @Override
+            public void addJwtsGenerationEntry(JwtsGenerationEntry jwtsGenerationEntry) {
+
+            }
+
+            @Override
+            public void addJwtsValidationEntry(JwtsValidationEntry jwtsValidationEntry) {
+
+            }
+
+            @Override
+            public void pushMetrics() {
+
+            }
+
+            @Override
+            public List<EndpointUsageEntry> getEndpointUsageMetric() {
+                return null;
+            }
+
+            @Override
+            public List<JwtsGenerationEntry> getJwtsGenerationMetric() {
+                return null;
+            }
+
+            @Override
+            public List<JwtsValidationEntry> getJwtsValidationMetric() {
+                return null;
+            }
+        };
+    }
+
+    @ConditionalOnMissingBean(PsuCounterEntryKPIService.class)
+    @Bean
+    PsuCounterEntryKPIService psuCounterEntryKPIService() {
+        return psuCounterEntryKPIService -> {};
+    }
+
+    @ConditionalOnMissingBean(SessionCountersKPIService.class)
+    @Bean
+    SessionCountersKPIService sessionCountersKPIService() {
+        return sessionCountersKPIService -> {};
+    }
+
+
+    @ConditionalOnMissingBean(TokenUsageService.class)
+    @Bean
+    TokenUsageService tokenUsageService() {
+        return  new TokenUsageService() {
+            @Override
+            public void incrementTokenUsage(TokenUsage... tokenUsage) {
+
+            }
+
+            @Override
+            public void incrementTokenUsage(AccessTokenResponse accessTokenResponse) {
+
+            }
+
+            @Override
+            public void increment(Collection<TokenUsage> tokenUsages) {
+
+            }
+        };
+    }
+
+
+    @ConditionalOnMissingBean(TppEntriesKPIService.class)
+    @Bean
+    TppEntriesKPIService tppEntriesKPIService() {
+        return tppEntriesKPIService -> {};
+    }
 
     @Override
     public String getIssuerID() {
