+++ forgerock-openbanking-analytics-client/src/main/java/com/forgerock/openbanking/analytics/services/CallBackCountersKPIService.java
@@ -20,48 +20,14 @@
  */
 package com.forgerock.openbanking.analytics.services;
 
-import brave.Tracer;
-import com.forgerock.openbanking.analytics.configuration.MetricsConfiguration;
 import com.forgerock.openbanking.analytics.model.entries.callback.CallBackCounterEntry;
-import com.forgerock.openbanking.analytics.utils.MetricUtils;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import org.springframework.web.reactive.function.BodyInserters;
-import org.springframework.web.reactive.function.client.WebClient;
 
 import java.util.List;
 
-@Slf4j
 @Service
-public class CallBackCountersKPIService {
+public interface CallBackCountersKPIService {
 
-    private final WebClient webClient;
-    private final MetricsConfiguration metricsConfiguration;
-    private final Tracer tracer;
+    public void sendCallBackEntries(List<CallBackCounterEntry> callBackCounterEntries);
 
-    public CallBackCountersKPIService(WebClient webClient, MetricsConfiguration metricsConfiguration, Tracer tracer) {
-        this.webClient = webClient;
-        this.metricsConfiguration = metricsConfiguration;
-        this.tracer = tracer;
-    }
-
-    public void sendCallBackEntries(List<CallBackCounterEntry> callBackCounterEntries) {
-        if (!MetricUtils.isRequestEnabledForAnalytics(tracer)) {
-            log.debug("Request excluded for analytics");
-            return;
-        }
-        log.debug("Send call back entries {}", callBackCounterEntries);
-        try {
-            webClient
-                    .post()
-                    .uri(metricsConfiguration.rootEndpoint + "/api/kpi/callbacks/")
-                    .body(BodyInserters.fromObject(callBackCounterEntries))
-                    .retrieve().bodyToMono(String.class)
-                    .log()
-                    .subscribe(response -> log.debug("Response from metrics: {}", response));
-            ;
-        } catch (Exception e) {
-            log.warn("Couldn't send callbacks {} KPI to metrics", callBackCounterEntries, e);
-        }
-    }
 }
