+++ forgerock-openbanking-analytics-client/src/main/java/com/forgerock/openbanking/analytics/services/PsuCounterEntryKPIService.java
@@ -20,45 +20,10 @@
  */
 package com.forgerock.openbanking.analytics.services;
 
-import brave.Tracer;
-import com.forgerock.openbanking.analytics.configuration.MetricsConfiguration;
 import com.forgerock.openbanking.analytics.model.entries.PsuCounterEntry;
-import com.forgerock.openbanking.analytics.utils.MetricUtils;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.web.reactive.function.BodyInserters;
-import org.springframework.web.reactive.function.client.WebClient;
 
-@Slf4j
-@Service
-public class PsuCounterEntryKPIService {
+public interface PsuCounterEntryKPIService {
 
-    private final WebClient webClient;
-    private final MetricsConfiguration metricsConfiguration;
-    private final Tracer tracer;
+    public void pushPsuCounterEntry(PsuCounterEntry entry);
 
-    public PsuCounterEntryKPIService(WebClient webClient, MetricsConfiguration metricsConfiguration, Tracer tracer) {
-        this.webClient = webClient;
-        this.metricsConfiguration = metricsConfiguration;
-        this.tracer = tracer;
-    }
-
-    public void pushPsuCounterEntry(PsuCounterEntry entry) {
-        if (!MetricUtils.isRequestEnabledForAnalytics(tracer)) {
-            log.debug("Request excluded for analytics");
-            return;
-        }
-        log.debug("Push psu counter entry={}", entry);
-        try {
-            webClient
-                    .post()
-                    .uri(metricsConfiguration.rootEndpoint + "/api/kpi/psu")
-                    .body(BodyInserters.fromObject(entry))
-                    .retrieve().bodyToMono(String.class)
-                    .log()
-                    .subscribe(response -> log.debug("Response from metrics: {}", response));
-        } catch (Exception e) {
-            log.warn("Couldn't send psu entry", entry, e);
-        }
-    }
 }
