+++ forgerock-openbanking-analytics-client/src/main/java/com/forgerock/openbanking/analytics/services/TppEntriesKPIService.java
@@ -20,47 +20,10 @@
  */
 package com.forgerock.openbanking.analytics.services;
 
-import brave.Tracer;
-import com.forgerock.openbanking.analytics.configuration.MetricsConfiguration;
 import com.forgerock.openbanking.analytics.model.entries.TppEntry;
-import com.forgerock.openbanking.analytics.utils.MetricUtils;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.web.reactive.function.BodyInserters;
-import org.springframework.web.reactive.function.client.WebClient;
 
-import java.util.Arrays;
+public interface TppEntriesKPIService {
 
-@Slf4j
-@Service
-public class TppEntriesKPIService {
+    void pushTppEntry(TppEntry entry);
 
-    private final WebClient webClient;
-    private final MetricsConfiguration metricsConfiguration;
-    private final Tracer tracer;
-
-    public TppEntriesKPIService(WebClient webClient, MetricsConfiguration metricsConfiguration, Tracer tracer) {
-        this.webClient = webClient;
-        this.metricsConfiguration = metricsConfiguration;
-        this.tracer = tracer;
-    }
-
-    public void pushTppEntry(TppEntry entry) {
-        if (!MetricUtils.isRequestEnabledForAnalytics(tracer)) {
-            log.debug("Request excluded for analytics");
-            return;
-        }
-        log.debug("Push tpp entry={}", entry);
-        try {
-            webClient
-                    .post()
-                    .uri(metricsConfiguration.rootEndpoint + "/api/kpi/tpps/entries")
-                    .body(BodyInserters.fromObject(Arrays.asList(entry)))
-                    .retrieve().bodyToMono(String.class)
-                    .log()
-                    .subscribe(response -> log.debug("Response from metrics: {}", response));
-        } catch (Exception e) {
-            log.warn("Couldn't send Tpp entry", entry, e);
-        }
-    }
 }
