--- forgerock-openbanking-analytics-client/src/test/resources/dev-transport.pem
@@ -1,27 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFcDCCBFigAwIBAgIUC4L3x9iIstlZ4mZHRFgRcEHTAeYwDQYJKoZIhvcNAQELBQAwezELMAkG
-A1UEBhMCVUsxDTALBgNVBAgTBEF2b24xEDAOBgNVBAcTB0JyaXN0b2wxEjAQBgNVBAoTCUZvcmdl
-Um9jazEcMBoGA1UECxMTZm9yZ2Vyb2NrLmZpbmFuY2lhbDEZMBcGA1UEAxMQb2JyaS1leHRlcm5h
-bC1jYTAgFw0xNzEyMDYxNTExMjJaGA8yMTE4MTExMjE1MTEyMlowgbExITAfBgNVBAMMGDVjMDkz
-ZmFhODQ2NGJlM2YwMmU2Nzg0ZjEhMB8GA1UECwwYNWJlMWEyYTVjNzdiMjMzNDI4NDI0M2JhMRIw
-EAYDVQQKDAlGb3JnZVJvY2sxEDAOBgNVBAcMB0JyaXN0b2wxDTALBgNVBAgMBEF2b24xCzAJBgNV
-BAYTAlVLMScwJQYDVQRhDB5QU0RHQi01YmUxYTJhNWM3N2IyMzM0Mjg0MjQzYmEwggEiMA0GCSqG
-SIb3DQEBAQUAA4IBDwAwggEKAoIBAQCDuMhPNcj1wF4BIi/8mMFH0Ik8yH/BOYJ0rXwaElPUIqQ7
-4agkYeJD6kIGC3HEB4WPJRoaN5SPEp8rHalD88lDLQvHLEv8Hky3gGD463zUTJCnDwTbSM3ndDSo
-AVp8SrPAsolgDOIfU+phgSeqJeKLwyb8bm3LJDZHeVYMZFmkS5KYYZDB5Ml0ZEgH4MFBs2i+LD/l
-yjUSAoLo2tjVvA62jWtZFKFDogMnCFhFLXoBQCXFcabwOfLLidrBPsCln2nDGcFXgLKwam/e5w/w
-3oNhoovp7kFc8Z66p/cUhdQol5D7/XljZg6l4iiAbX/MXDRUWiKGpPSpbJL6yBdOLCSRAgMBAAGj
-ggGxMIIBrTCB1AYIKwYBBQUHAQEEgccwgcQwYAYIKwYBBQUHMAKGVGh0dHBzOi8vc2VydmljZS5k
-aXJlY3RvcnkuZGV2LW9iLmZvcmdlcm9jay5maW5hbmNpYWw6ODA3NC9hcGkvZGlyZWN0b3J5L2tl
-eXMvandrX3VyaTBgBggrBgEFBQcwAYZUaHR0cHM6Ly9zZXJ2aWNlLmRpcmVjdG9yeS5kZXYtb2Iu
-Zm9yZ2Vyb2NrLmZpbmFuY2lhbDo4MDc0L2FwaS9kaXJlY3Rvcnkva2V5cy9qd2tfdXJpMIHTBggr
-BgEFBQcBAwSBxjCBwzAIBgYEAI5GAQEwCQYHBACORgEGAzAJBgcEAIvsSQECMIGgBgYEAIGYJwIw
-gZUwajApBgcEAIGYJwEEDB5DYXJkIEJhc2VkIFBheW1lbnQgSW5zdHJ1bWVudHMwHgYHBACBmCcB
-AwwTQWNjb3VudCBJbmZvcm1hdGlvbjAdBgcEAIGYJwECDBJQYXltZW50IEluaXRpYXRpb24MHUZv
-cmdlUm9jayBGaW5hbmNpYWwgQXV0aG9yaXR5DAhGUi1BQUFBQTANBgkqhkiG9w0BAQsFAAOCAQEA
-k03mFp/eDhvQDibSgWXZLqobZ8GtVzLulRnardVRlDruinrr4QJUANSCtCvImUqx2HAWtZ3o5FsV
-pg7soEjzYwfYGgO2AfrRF+6u7HBs82BZWu8EYa62P7SKg+tJRfUwGNLIsgNAlNp8r1HRj+5KrFla
-pAr++QZW7uiJMQ4jvo4OuMkTPBMq9dn7L+WsSkNMluZE6nXluWL4Mj3TR6HKh9ECq1+Fdu/LiXcF
-HGYVVIxy0Vo5vMp55xWVkByiLfzYJDgo6+YUM3Na38YLFTJAaqZFq5EMRy7JMnvjhdbZFhknr0+N
-ERWUTMpP4xpUTuiSOmZVlet5ENq7tanURmC3kA==
------END CERTIFICATE-----
