--- forgerock-openbanking-analytics-client/src/test/resources/fr-directory.pem
@@ -1,19 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDmDCCAoCgAwIBAgIUfGwulNQG7ZAlpIurC09frE64igMwDQYJKoZIhvcNAQELBQAwezELMAkG
-A1UEBhMCVUsxDTALBgNVBAgTBEF2b24xEDAOBgNVBAcTB0JyaXN0b2wxEjAQBgNVBAoTCUZvcmdl
-Um9jazEcMBoGA1UECxMTZm9yZ2Vyb2NrLmZpbmFuY2lhbDEZMBcGA1UEAxMQb2JyaS1leHRlcm5h
-bC1jYTAgFw0xNzExMjIyMjI2MDBaGA8yMTE4MTAyOTIyMjYwMFowgY4xCzAJBgNVBAYTAlVLMQ0w
-CwYDVQQIEwRBdm9uMRAwDgYDVQQHEwdCcmlzdG9sMRIwEAYDVQQKEwlGb3JnZVJvY2sxHDAaBgNV
-BAsTE2Zvcmdlcm9jay5maW5hbmNpYWwxLDAqBgNVBAMTI3RwcC1jb3JlLmRldi1vYi5mb3JnZXJv
-Y2suZmluYW5jaWFsMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAnrce8EwCMzbm1Wa5
-Xan9TkLskOE3LG8wQK4hpq987t+q/L4azM2wkf/Lj096rKYX36d3+vt3GJYqFL0YV6Z56Uy/Dot0
-RrBs2NtvAlvle+7+zRj+9/A/N+mCIrqHoJytWY5PjAZ8yfP2McsyxpICMm7jehFiI+3PbvgCz+V3
-qR9hii1nfNuTr1iyau5ENriaopNMhwKRDjIDFtbvDvDj84btk/6WDfqflrEKAJ8nLBH9rE6Y2MN/
-sY5ZTdfyl+f9D2pvF56iJvQeH9YSkRZNZpE2DWUUKovBaOaj9fojcvoYkPn++MmXZch36CvGjLej
-NKBcY0IeFS1lyh5ndfL+mwIDAQABMA0GCSqGSIb3DQEBCwUAA4IBAQAZ2f07G6UlNt/FICXDdqOD
-2IyPFWhOFHRr59O5+zbeWLA7hPXkVQY0eBSBxNdTKIE/P5dpiMkLvqpONcs8GF5y2srUPxKxXx8Q
-PNMWp3cHGPobs2WdYUxe2z5rS8y79y0bblTG1vHhzuH8ndDrn/6Uu3iLiRbf2PRgLzXB9ZL49Bti
-us/qq1/MpigL0lvbvYs/BdkXaQI7fD8HU11Neegd5bat364JMsYBKtQr38Ai2JQzA7URkAfcBaLT
-TZLbMMcmzXbZXC+/wqtvOLvLv7rpl334BT2eOFaOwblhstbU0hhg+8utHeX+w/wyv2useoaEW8ru
-OyKRf9C80jMjnvqp
------END CERTIFICATE-----
