--- forgerock-openbanking-analytics-client/src/test/resources/payment/OBIEPaymentInitiation_3_0.json
@@ -1,78 +1,0 @@
-{
-  "Data": {
-    "DomesticPayments": [{
-      "InstructionIdentification": "ANSM020",
-      "EndToEndIdentification": "FRESCO.21302.GFX.01",
-      "LocalInstrument": "UK.OBIE.CHAPS",
-      "InstructedAmount": {
-        "Amount": "21.00",
-        "Currency": "GBP"
-      },
-      "DebtorAccount": {
-        "SchemeName": "UK.OBIE.SortCodeAccountNumber",
-        "Identification": "11280001234567",
-        "Name": "Andrea Smith"
-      },
-      "CreditorAccount": {
-        "SchemeName": "UK.OBIE.SortCodeAccountNumber",
-        "Identification": "08080021325698",
-        "Name": "Bob Clements"
-      },
-      "CreditorPostalAddress": {
-        "AddressType": "Correspondence",
-        "StreetName": "Liberty",
-        "BuildingNumber": "1",
-        "PostCode": "AB1 2CD",
-        "TownName": "London",
-        "Country": "UK"
-      },
-      "RemittanceInformation": {
-        "Reference": "FRESCO-037",
-        "Unstructured": "Internal ops code 5120103"
-      }
-    }, {
-      "InstructionIdentification": "ANSM021",
-      "EndToEndIdentification": "FRESCO.21302.GFX.02",
-      "LocalInstrument": "UK.OBIE.BACS",
-      "InstructedAmount": {
-        "Amount": "22.00",
-        "Currency": "GBP"
-      },
-      "DebtorAccount": {
-        "SchemeName": "UK.OBIE.SortCodeAccountNumber",
-        "Identification": "11280001234567",
-        "Name": "Andrea Smith"
-      },
-      "CreditorAccount": {
-        "SchemeName": "UK.OBIE.SortCodeAccountNumber",
-        "Identification": "08080021325698",
-        "Name": "Bob Clements"
-      },
-      "RemittanceInformation": {
-        "Reference": "FRESCO-037",
-        "Unstructured": "Internal ops code 5120103"
-      }
-    }, {
-      "InstructionIdentification": "ANSM022",
-      "EndToEndIdentification": "FRESCO.21302.GFX.03",
-      "InstructedAmount": {
-        "Amount": "23.00",
-        "Currency": "GBP"
-      },
-      "DebtorAccount": {
-        "SchemeName": "SortCodeAccountNumber",
-        "Identification": "11280001234567",
-        "Name": "Andrea Smith"
-      },
-      "CreditorAccount": {
-        "SchemeName": "SortCodeAccountNumber",
-        "Identification": "08080021325698",
-        "Name": "Bob Clements"
-      },
-      "RemittanceInformation": {
-        "Reference": "FRESCO-037",
-        "Unstructured": "Internal ops code 5120103"
-      }
-    }]
-  }
-}
