--- forgerock-openbanking-analytics-client/src/test/resources/x509.pem
@@ -1,13 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIB5jCCAZACCQDaek5z5yfbLDANBgkqhkiG9w0BAQsFADB6MQswCQYDVQQGEwJ1
-czELMAkGA1UECAwCbnkxCzAJBgNVBAcMAm55MQswCQYDVQQKDAJmcjELMAkGA1UE
-CwwCb2IxFjAUBgNVBAMMDWZvcmdlcm9jay5jb20xHzAdBgkqhkiG9w0BCQEWEG9i
-QGZvcmdlcm9jay5jb20wHhcNMTgxMDEwMTMzOTEzWhcNMTkxMDEwMTMzOTEzWjB6
-MQswCQYDVQQGEwJ1czELMAkGA1UECAwCbnkxCzAJBgNVBAcMAm55MQswCQYDVQQK
-DAJmcjELMAkGA1UECwwCb2IxFjAUBgNVBAMMDWZvcmdlcm9jay5jb20xHzAdBgkq
-hkiG9w0BCQEWEG9iQGZvcmdlcm9jay5jb20wXDANBgkqhkiG9w0BAQEFAANLADBI
-AkEAyeiYqTr3IxRRr/ZniUzM7AeaIcADtapAfKPqfG0TSzjAp3/kVCuov7CTldAQ
-YogSZx/vXgEHEOJF7eju9fWkvwIDAQABMA0GCSqGSIb3DQEBCwUAA0EAOdytdCAS
-QN7AHYu7HWwRpLcc7c8mYKVtUwEGN+hR9SuNTB2XmYCht2Trs4vWyRPCP+pKMkkY
-FbbVxupW5TcZaQ==
------END CERTIFICATE-----
