+++ forgerock-openbanking-analytics-webclient/src/main/java/com/forgerock/openbanking/analytics/webclient/services/WebClientConsentMetricService.java
@@ -1,0 +1,68 @@
+/**
+ * Copyright 2019 ForgeRock AS.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package com.forgerock.openbanking.analytics.webclient.services;
+
+import brave.Tracer;
+import com.forgerock.openbanking.analytics.model.entries.ConsentStatusEntry;
+import com.forgerock.openbanking.analytics.services.ConsentMetricService;
+import com.forgerock.openbanking.analytics.webclient.configuration.MetricsConfiguration;
+import com.forgerock.openbanking.analytics.utils.MetricUtils;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestClientException;
+import org.springframework.web.reactive.function.BodyInserters;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@Service
+@Slf4j
+public class WebClientConsentMetricService implements ConsentMetricService {
+
+    private final WebClient webClient;
+    private final MetricsConfiguration metricsConfiguration;
+    private final Tracer tracer;
+
+
+    public WebClientConsentMetricService(WebClient webClient, MetricsConfiguration metricsConfiguration, Tracer tracer) {
+        this.webClient = webClient;
+        this.metricsConfiguration = metricsConfiguration;
+        this.tracer = tracer;
+    }
+
+    public void sendConsentActivity(ConsentStatusEntry consentStatusEntry) {
+        if (!MetricUtils.isRequestEnabledForAnalytics(tracer)) {
+            log.debug("Request excluded for analytics");
+            return;
+        }
+        log.debug("Incrementing consent activity: {}", consentStatusEntry);
+        try {
+            webClient
+                    .post()
+                    .uri(metricsConfiguration.rootEndpoint + "/api/kpi/consent")
+                    .body(BodyInserters.fromObject(consentStatusEntry))
+                    .retrieve().bodyToMono(String.class)
+                    .log()
+                    .subscribe(response -> log.debug("Response from metrics: {}", response));
+            ;
+        } catch (RestClientException e) {
+            log.warn("Could not update consent status metrics type={}", consentStatusEntry, e);
+        }
+    }
+}
